support secure read-only mode and query/exec on transactions
opening a db with ?mode=ro is not enough to make it secure - e.g. attach can
still be used to attach to other databases and thus read data that was not
meant to be read. Rather, we have to make use of the authorizer interface and
deny everything but selecting/reading/functions.