Add json_includes sql function
As we want to be generic we abuse the string representation of the type. We
could just as well convert both to json and compare the json strings but this
one is easier and good enough for me (for now). We can't just compare as is
because numeric types can differ.