Add naive repl
The custom repl provides access to gosql functions (not that many, but still)
and can print output a little prettier (aligned columns). Also limiting column
length to 100 because it's simple and fits my current needs. Should be replace
with `.width x y z` later on. The whole tabwriting to string first is so we can
ctr-c - i care more about ctrl-c than performance.